Fix object-dtype plotting bug, add max_epochs param, save before diagnostics
- Fix plot_training_diagnostics hanging: model.history_ DataFrames have
  dtype('O') from Lightning; add .astype(float) to all .values.ravel() calls
  in both _model.py and _multimodel.py
- Multimodal notebook: swap model save before diagnostics plot to prevent
  losing trained model if plotting hangs
- Add max_epochs as papermill parameter (default 5000) for dry-run testing
- Add skip_training papermill parameter to load saved model without retraining
- Add use_feature_scaling papermill parameter for ablation experiments
- Make skip_epochs adaptive: min(80, n_epochs // 2)
- Add _gpu_jobs.yaml with explicit notebook paths and detailed descriptions
- Add _test_diagnostics.py helper for inspecting saved model history

+++ src/regularizedvi/_model.py
@@ -857,12 +857,12 @@ class AmbientRegularizedSCVI(
 
             if train_key is not None and train_key in history:
                 df = history[train_key]
-                values = df.iloc[skip_epochs:].values.ravel()
+                values = df.iloc[skip_epochs:].values.ravel().astype(float)
                 epochs = range(skip_epochs, skip_epochs + len(values))
                 ax.plot(epochs, values, color="tab:blue", label="train")
             if val_key is not None and val_key in history:
                 df = history[val_key]
-                values = df.iloc[skip_epochs:].values.ravel()
+                values = df.iloc[skip_epochs:].values.ravel().astype(float)
                 epochs = range(skip_epochs, skip_epochs + len(values))
                 ax.plot(epochs, values, color="tab:orange", label="validation")
